+++ lib/errors.rb
@@ -1,0 +1,3 @@
+class ArgumentTypeError < TypeError; end;
+class LocalTypeError < TypeError; end;
+class ReturnTypeError < TypeError; end;