+++ lib/interfaces/proxy_interface.rb
@@ -1,0 +1,11 @@
+module LowType
+  class ProxyInterface
+    def error_type(value:)
+      raise NotImplementedError
+    end
+
+    def error_message(value:, line: nil)
+      raise NotImplementedError
+    end
+  end
+end