+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.7.0
+
+### Added
+
+- Raise `ArgumentTypeError`, `LocalTypeError` and `ReturnTypeError` error types
+- Improve error messages
+
 ## 0.6.0
 
 - `type` and `object.with_type=()` assignment methods